feat(container): add query params to sandbox links
Add a query param to the CodeSandbox link in PixiJSContainer component that sets the default file to Demo.vue. Add a comment to disable eslint in StackBlitzEmbed component and include a query param that sets the default file to App.vue in the options object passed to the StackBlitz sdk.