feat: Add Loader and AnimatedSprite props changes
This commit adds changes to props of Loader and AnimatedSprite components. It updates the type of the `onResolved` prop in the Loader component and adds `roundPixels` prop to the AnimatedSprite component. It also removes the `textures` prop from the set of props handled by `patchTextureProps` and adds it to `patchAnimatedSpriteProps`. The latter now uses the `normalizeTexture` function to update this prop and also resets the sprite's animation to frame 0 if the `loop` property is true.