feat(elements): add SimplePlane element and demo
Adds the SimplePlane element to the Elements section of the documentation, along with a basic demo of the element. The SimplePlane allows manipulation of texture points across several points, and its demo showcases this with randomized vertice positions to create movement. Also includes a minor style change to hide scrollbars within code blocks.