feat(api): add blend mode to Mesh and ParticleContainer
Added `blend-mode` attribute to `Mesh` and `ParticleContainer` components in order to the user to specify the blend mode to be applied to the sprite. Also added `ApiBendMode.md` to document the `BLEND_MODES` enum used for this attribute. Updated the documentation for the `NineSlicePlane` and `AnimatedSprite` components to use the `<api />` custom component for the `blend` attribute.